refactor: merge EEVM.Receipt into EEVM.Block.Receipt
Two receipt modules existed side by side: EEVM.Block.Receipt held the
struct and constructors, while EEVM.Receipt re-defined the same struct
plus RLP/EIP-2718 encoding. Only the test suite ever pulled in the
top-level module, so the duplication was dead weight that risked future
divergence.

Absorb the encode/2 path and its helpers into EEVM.Block.Receipt and
delete EEVM.Receipt. Move the test alongside its module under
test/block/.
