+++ lib/eevm/block/receipt.ex
@@ -20,22 +20,22 @@ defmodule EEVM.Block.Receipt do
     `EEVM.Block.Bloom.from_logs(logs)`. We store it pre-computed so block-level
     aggregation is a fast bytewise OR.
 
-  The processor fills in `cumulative_gas_used` and `logs_bloom` after each
-  transaction; the injected tx executor is only responsible for the raw
-  status / gas / logs triple.
+  ## Wire encoding (EIP-2718)
 
-  ## Elixir Learning Notes
+  - **Legacy** (type-0) receipts encode as a plain RLP list:
+    `RLP([status, cumulative_gas_used, logs_bloom, logs])`.
+  - **Typed** receipts (EIP-2930 / 1559 / 4844) prepend a single type byte to
+    the same RLP list, e.g. `<<0x02>> <> RLP([...])` for EIP-1559.
 
-  - This is a thin struct — no logic. Constructors default to a successful
-    receipt with no logs so tests can build interesting cases with just the
-    overrides that matter.
-  - The `logs` type mirrors `EEVM.Block.Bloom.log_entry/0` exactly to keep the
-    two modules compatible without mutual imports.
+  Each log encodes as `[address, [topic, ...], data]`, with the address as a
+  20-byte big-endian binary and each topic as a 32-byte big-endian binary.
   """
 
   alias EEVM.Block.Bloom
 
   @type log :: Bloom.log_entry()
+
+  @type tx_type :: :legacy | :eip2930 | :eip1559 | :eip4844 | non_neg_integer()
 
   @type t :: %__MODULE__{
           status: 0 | 1,
@@ -48,6 +48,10 @@ defmodule EEVM.Block.Receipt do
             cumulative_gas_used: 0,
             logs: [],
             logs_bloom: <<0::2048>>
+
+  @type_access_list 0x01
+  @type_fee_market 0x02
+  @type_blob 0x03
 
   @doc "Returns a successful receipt with no logs and zero cumulative gas."
   @spec new() :: t()
@@ -70,4 +74,50 @@ defmodule EEVM.Block.Receipt do
       true -> base
     end
   end
+
+  @doc """
+  Encode a receipt to its wire (RLP) form.
+
+  Pass `tx_type:` to produce a typed (EIP-2718) receipt. The accepted values
+  are `:legacy` (default), `:eip2930`, `:eip1559`, `:eip4844`, or the raw
+  integer type byte.
+  """
+  @spec encode(t(), keyword()) :: binary()
+  def encode(%__MODULE__{} = receipt, opts \\ []) do
+    payload = ExRLP.encode(rlp_fields(receipt))
+
+    case type_byte(Keyword.get(opts, :tx_type, :legacy)) do
+      nil -> payload
+      type_byte -> <<type_byte>> <> payload
+    end
+  end
+
+  defp rlp_fields(%__MODULE__{} = receipt) do
+    [
+      encode_integer(receipt.status),
+      encode_integer(receipt.cumulative_gas_used),
+      receipt.logs_bloom,
+      Enum.map(receipt.logs, &encode_log/1)
+    ]
+  end
+
+  defp encode_log(%{address: address, topics: topics, data: data})
+       when is_integer(address) and is_list(topics) and is_binary(data) do
+    [
+      <<address::unsigned-big-160>>,
+      Enum.map(topics, fn topic -> <<topic::unsigned-big-256>> end),
+      data
+    ]
+  end
+
+  defp encode_integer(0), do: <<>>
+
+  defp encode_integer(value) when is_integer(value) and value > 0,
+    do: :binary.encode_unsigned(value)
+
+  defp type_byte(:legacy), do: nil
+  defp type_byte(:eip2930), do: @type_access_list
+  defp type_byte(:eip1559), do: @type_fee_market
+  defp type_byte(:eip4844), do: @type_blob
+  defp type_byte(byte) when is_integer(byte) and byte >= 0 and byte <= 0x7F, do: byte
 end
